+++ src/scenes/room_gen/init.rs
@@ -5,7 +5,7 @@ use super::util::pop_random;
 use crate::tuple_abs_diff;
 
 pub fn init_rects(size: usize, split_limit: usize) -> Vec<Vec<Objects>> {
-    let mut room = vec![vec![Objects::Air; size as usize]; size as usize];
+    let mut room = vec![vec![Objects::Air; size]; size];
 
     let mut queue = vec![SplitQueue::new(
         if gen_range(0, 2) == 0 {
